# perf(GQL): RHICOMPL-3176 preload benchmarks under profiles