--- a/test/graphql/queries/profile_query_test.rb
+++ b/test/graphql/queries/profile_query_test.rb
@@ -85,6 +85,9 @@ class ProfileQueryTest < ActiveSupport::TestCase
               id
               refId
               ssgVersion
+              benchmark {
+                refId
+              }
           }
       }
     GRAPHQL