**Before:**
```
Completed 200 OK in 723ms (Views: 0.9ms | ActiveRecord: 25.0ms | Allocations: 638229)
```
**After:**
```
Completed 200 OK in 610ms (Views: 0.9ms | ActiveRecord: 22.9ms | Allocations: 500235)
```
With the query:
```bash
time curl -k -H "Content-Type: application/json" -H "X-RH-IDENTITY: $RHI" -XPOST -d '{
  "operationName": "supportedProfilesByOSMajor",
  "variables": {},
  "query": "query supportedProfilesByOSMajor {
      osMajorVersions {
          edges {
              node {
                  osMajorVersion
                  profiles {
                      id
                      name
                      refId
                      description
                      complianceThreshold
                      supportedOsVersions
                      benchmark {
                          id
                          refId
                          osMajorVersion
                      }
                  }
              }
          }
      }
      profiles(search: \"external = false and canonical = false\") {
          edges {
              node {
                  id
                  refId
                  benchmark {
                      refId
                  }
              }
          }
      }
  }"
}' "http://localhost:3000/api/compliance/graphql" | jq
```

Should help with other queries that involve benchmarks under profiles as well :zap: 

## Secure Coding Practices Checklist GitHub Link
- https://github.com/RedHatInsights/secure-coding-checklist

## Secure Coding Checklist
- [ ] Input Validation
- [ ] Output Encoding
- [ ] Authentication and Password Management
- [ ] Session Management
- [ ] Access Control
- [ ] Cryptographic Practices
- [ ] Error Handling and Logging
- [ ] Data Protection
- [ ] Communication Security
- [ ] System Configuration
- [ ] Database Security
- [ ] File Management
- [ ] Memory Management
- [x] General Coding Practices